+++ cmd/workflow-engine/cli/cli.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"log/slog"
 	"os"
+	"strings"
 	"workflow-engine/pkg/pipelines"
 
 	"github.com/spf13/cobra"
@@ -52,7 +53,7 @@ func (a *App) Init() {
 		Use:   "image-build",
 		Short: "Build a container image",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			if err := a.loadConfig(cmd, args); err != nil {
+			if err := a.loadConfig(); err != nil {
 				return err
 			}
 			return imageBuildPipeline(cmd, a.flagDryRun, imageBuildCmd(*a.flagCLICmd), a.cfg.Image)
@@ -63,7 +64,7 @@ func (a *App) Init() {
 		Use:   "image-scan",
 		Short: "Scan a container image",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			if err := a.loadConfig(cmd, args); err != nil {
+			if err := a.loadConfig(); err != nil {
 				return err
 			}
 			return imageScanPipeline(cmd, a.flagDryRun, a.cfg.Artifacts, a.cfg.Image.ScanTarget)
@@ -158,43 +159,43 @@ func (a *App) Init() {
 
 	// Image Build Bindings
 	//   Viper bind config keys to flag values and environment variables
-	viper.BindPFlag("buildDir", imagebuildCmd.Flags().Lookup("build-dir"))
-	viper.MustBindEnv("buildDir", "WFE_BUILD_DIR")
+	viper.BindPFlag("image.builddir", imagebuildCmd.Flags().Lookup("build-dir"))
+	viper.MustBindEnv("image.builddir", "WFE_BUILD_DIR")
 
-	viper.BindPFlag("buildDockerfile", imagebuildCmd.Flags().Lookup("dockerfile"))
-	viper.MustBindEnv("buildDockerfile", "WFE_BUILD_DOCKERFILE")
+	viper.BindPFlag("image.builddockerfile", imagebuildCmd.Flags().Lookup("dockerfile"))
+	viper.MustBindEnv("image.builddockerfile", "WFE_BUILD_DOCKERFILE")
 
-	viper.BindPFlag("buildTag", imagebuildCmd.Flags().Lookup("tag"))
-	viper.MustBindEnv("buildTag", "WFE_BUILD_TAG")
+	viper.BindPFlag("image.buildtag", imagebuildCmd.Flags().Lookup("tag"))
+	viper.MustBindEnv("image.buildtag", "WFE_BUILD_TAG")
 
-	viper.BindPFlag("buildPlatform", imagebuildCmd.Flags().Lookup("platform"))
-	viper.MustBindEnv("buildPlatform", "WFE_BUILD_PLATFORM")
+	viper.BindPFlag("image.buildplatform", imagebuildCmd.Flags().Lookup("platform"))
+	viper.MustBindEnv("image.buildplatform", "WFE_BUILD_PLATFORM")
 
-	viper.BindPFlag("buildTarget", imagebuildCmd.Flags().Lookup("target"))
-	viper.MustBindEnv("buildTarget", "WFE_BUILD_TARGET")
+	viper.BindPFlag("image.buildtarget", imagebuildCmd.Flags().Lookup("target"))
+	viper.MustBindEnv("image.buildtarget", "WFE_BUILD_TARGET")
 
-	viper.BindPFlag("buildCacheTo", imagebuildCmd.Flags().Lookup("cache-to"))
-	viper.MustBindEnv("buildCacheTo", "WFE_BUILD_CACHE_TO")
+	viper.BindPFlag("image.buildcacheto", imagebuildCmd.Flags().Lookup("cache-to"))
+	viper.MustBindEnv("image.buildcacheto", "WFE_BUILD_CACHE_TO")
 
-	viper.BindPFlag("buildCacheFrom", imagebuildCmd.Flags().Lookup("cache-from"))
-	viper.MustBindEnv("buildCacheFrom", "WFE_BUILD_CACHE_FROM")
+	viper.BindPFlag("image.buildcachefrom", imagebuildCmd.Flags().Lookup("cache-from"))
+	viper.MustBindEnv("image.buildcachefrom", "WFE_BUILD_CACHE_FROM")
 
-	viper.BindPFlag("buildSquashLayers", imagebuildCmd.Flags().Lookup("squash-layers"))
-	viper.MustBindEnv("buildSquashLayers", "WFE_BUILD_SQUASH_LAYERS")
+	viper.BindPFlag("image.buildsquashlayers", imagebuildCmd.Flags().Lookup("squash-layers"))
+	viper.MustBindEnv("image.buildsquashlayers", "WFE_BUILD_SQUASH_LAYERS")
 
 	// Image Scan Bindings
-	viper.BindPFlag("artifactDirectory", imageScanCmd.Flags().Lookup("artifact-directory"))
-	viper.MustBindEnv("artifactDirectory", "WFE_ARTIFACT_DIRECTORY")
+	viper.BindPFlag("artifacts.directory", imageScanCmd.Flags().Lookup("artifact-directory"))
+	viper.MustBindEnv("artifacts.directory", "WFE_ARTIFACT_DIRECTORY")
 
-	viper.BindPFlag("sbomFilename", imageScanCmd.Flags().Lookup("sbom-filename"))
-	viper.MustBindEnv("sbomFilename", "WFE_SBOM_FILENAME")
+	viper.BindPFlag("artifacts.sbomfilename", imageScanCmd.Flags().Lookup("sbom-filename"))
+	viper.MustBindEnv("artifacts.sbomfilename", "WFE_SBOM_FILENAME")
 
-	viper.BindPFlag("grypeFilename", imageScanCmd.Flags().Lookup("grype-filename"))
-	viper.MustBindEnv("grypeFilename", "WFE_GRYPE_FILENAME")
+	viper.BindPFlag("artifacts.grypefilename", imageScanCmd.Flags().Lookup("grype-filename"))
+	viper.MustBindEnv("artifacts.grypefilename", "WFE_GRYPE_FILENAME")
 
 	// TODO: need to consider the logic for overthe build tag here
-	viper.BindPFlag("scanImageTarget", imageScanCmd.Flags().Lookup("scan-image-target"))
-	viper.MustBindEnv("scanImageTarget", "WFE_SCAN_IMAGE_TARGET")
+	viper.BindPFlag("image.scantarget", imageScanCmd.Flags().Lookup("scan-image-target"))
+	viper.MustBindEnv("image.scantarget", "WFE_SCAN_IMAGE_TARGET")
 
 	a.cmd.AddCommand(runCmd, configCmd)
 }
@@ -204,14 +205,14 @@ func (a *App) Execute() error {
 	return a.cmd.Execute()
 }
 
-func (a *App) loadConfig(cmd *cobra.Command, args []string) error {
+func (a *App) loadConfig() error {
 	l := slog.Default().With("step", "load_config")
 
 	l.Debug("check config file flag value")
 	configFile, _ := a.cmd.PersistentFlags().GetString("config")
 	if configFile != "" {
 		viper.SetConfigFile(configFile)
-		return nil
+		l.Debug("viper config file set", "config_file", configFile)
 	}
 
 	// viper reads in config values from all sources based on precedence
@@ -226,24 +227,30 @@ func (a *App) loadConfig(cmd *cobra.Command, args []string) error {
 		}
 	}
 
+	var sb strings.Builder
+	for _, key := range viper.AllKeys() {
+		slog.Debug("config", "key", key, "value", fmt.Sprintf("%v", viper.Get(key)))
+	}
+	slog.Debug(sb.String())
+
 	// viper will unmarshal the values into the cfg object
 	l.Debug("decode configuration file")
 	a.cfg = &pipelines.Config{
 		Image: pipelines.ImageConfig{
-			BuildDir:        viper.GetString("buildDir"),
-			BuildDockerfile: viper.GetString("buildDockerfile"),
-			BuildTag:        viper.GetString("buildTag"),
-			BuildPlatform:   viper.GetString("buildPlatform"),
-			BuildTarget:     viper.GetString("buildTarget"),
-			BuildCacheTo:    viper.GetString("buildCacheTo"),
-			BuildCacheFrom:  viper.GetString("buildCacheFrom"),
+			BuildDir:        viper.GetString("image.builddir"),
+			BuildDockerfile: viper.GetString("image.builddockerfile"),
+			BuildTag:        viper.GetString("image.buildtag"),
+			BuildPlatform:   viper.GetString("image.buildplatform"),
+			BuildTarget:     viper.GetString("image.buildtarget"),
+			BuildCacheTo:    viper.GetString("image.buildcacheto"),
+			BuildCacheFrom:  viper.GetString("image.buildcachefrom"),
 			BuildArgs:       make([][2]string, 0),
-			ScanTarget:      viper.GetString("scanImageTarget"),
+			ScanTarget:      viper.GetString("image.scantarget"),
 		},
 		Artifacts: pipelines.ArtifactConfig{
-			Directory:     viper.GetString("artifactDirectory"),
-			SBOMFilename:  viper.GetString("sbomFilename"),
-			GrypeFilename: viper.GetString("grypeFilename"),
+			Directory:     viper.GetString("artifacts.directory"),
+			SBOMFilename:  viper.GetString("artifacts.sbomfilename"),
+			GrypeFilename: viper.GetString("artifacts.grypefilename"),
 		},
 	}